requirements.txt
aiogram==2.25.1
aiohappyeyeballs==2.3.5
aiohttp==3.8.6
aiosignal==1.3.1
annotated-types==0.7.0
async-timeout==4.0.3
attrs==24.2.0
Babel==2.9.1
beautifulsoup4==4.11.2
certifi==2024.7.4
charset-normalizer==3.3.2
frozenlist==1.4.1
idna==3.7
magic-filter==1.0.12
multidict==6.0.5
pillow==10.4.0
pydantic==2.8.2
pydantic_core==2.20.1
python-dotenv==0.21.1
pytz==2024.1
requests==2.28.2
soupsieve==2.6
typing_extensions==4.12.2
urllib3==1.26.19
yarl==1.9.4
